requirements.txt
numpy
PyOpenGL
glfw
imgui[glfw]
tifffile
Pillow
PyGLM
PyQt6
